Allow using empty openstack_domain
Deployers need the ability to opt-out of RFC1034/5 hostnames. This is desirable in an upgrade scenario where short hostnames are already in use, because dropping the RFC1034/5 hostname in /etc/hosts during upgrade will cause the nova-compute service to remove the original row from the nova.compute_nodes table and create a new row with the new hostname. This breaks hypervisor stats reporting and potentially other operations (e.g., migration) since the nova.instances.node value refers to the original short hostname for instances already launched on the host.

This change simply moves the hard-coded dot (.) from the template for the hostname script into the openstack_domain variable, which should cause no change in behavior for the default scenario. However, it allows openstack_domain to be overridden with an empty string to avoid adding a domain to the short hostname altogether. Without this change, overriding openstack_domain with an empty string results in host entries with a tailing dot which is an invalid format and breaks some services (e.g., cinder).